+++ tests/CloudTek.Sdk.Tests/projects/OutdatedSdk.Error/Program.cs
@@ -1,0 +1,2 @@
+// See https://aka.ms/new-console-template for more information
+Console.WriteLine("Hello, World!");
