+++ tests/CloudTek.Sdk.Tests/projects/OutdatedSdk.Error/global.json
@@ -1,0 +1,6 @@
+{
+    "sdk": {
+      "version": "6.0.100",
+      "rollForward": "latestFeature"
+    }
+  }
