+++ tests/CloudTek.Sdk.Tests/projects/OutdatedSdk.Error/message.txt
@@ -1,0 +1,1 @@
+CloudTek.Sdk requires .NET SDK v8.0.100+ to be used
